+++ src/expressions/include/antares/expressions/visitors/TimeIndex.h
@@ -1,4 +1,3 @@
-
 /*
 ** Copyright 2007-2024, RTE (https://www.rte-france.com)
 ** See AUTHORS.txt
@@ -20,13 +19,39 @@
 ** along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
 */
 #pragma once
-#include <stdexcept>
 
-namespace Antares::Solver::Visitors
+namespace Antares::Expressions::Visitors
 {
-class InvalidNode: public std::invalid_argument
+/**
+ * @brief Represents the time and scenario variation of a value.
+ */
+enum class TimeIndex : unsigned int
 {
-public:
-    explicit InvalidNode(const std::string& node_name = "");
+    CONSTANT_IN_TIME_AND_SCENARIO = 0,
+    VARYING_IN_TIME_ONLY = 1,
+    VARYING_IN_SCENARIO_ONLY = 2,
+    VARYING_IN_TIME_AND_SCENARIO = 3
 };
-} // namespace Antares::Solver::Visitors
+
+/**
+ * @brief Combines two TimeIndex values.
+ *
+ * @param left The left operand.
+ * @param right The right operand.
+ *
+ * @return The combined TimeIndex value.
+ */
+constexpr TimeIndex operator|(const TimeIndex& left, const TimeIndex& right)
+{
+    /*
+     0 | x = x
+     3 | x = 3
+     1 | 1 = 1
+     1 | 2 = 3
+     2 | 2 = 2
+     */
+    return static_cast<TimeIndex>(static_cast<unsigned int>(left)
+                                  | static_cast<unsigned int>(right));
+}
+
+} // namespace Antares::Expressions::Visitors
